feat: add runtime lib.

+++ tigerc/src/data_layout.rs
@@ -1,7 +1,7 @@
 // |8 byte for length| data |
-pub const ARRAY_HEADER_SIZE: i64 = 8;
+pub const ARRAY_HEADER_SIZE: usize = 8;
 
 // |8 byte for string length| string | null |
 // compatible with c-string by null-terminated
 // so length of data is string length + 1
-pub const STRING_HEADER_SIZE: i64 = 8;
+pub const STRING_HEADER_SIZE: usize = 8;
